# dist/swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Universal Directory API",
        "version": "1.0.0",
        "description": ""
    },
    "servers": [
        {
            "url": "https://oauth.igoodworks.com/"
        }
    ],
    "paths": {
...
        "/jwks.json": {
            "options": {
                "description": "",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "get": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api-docs": {
            "get": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "oauth/v1/signup": {
            "post": {
                "description": "",
                "parameters": [],
                "responses": {
                    "default": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "oauth/v1/token": {
            "post": {
                "description": "",
                "parameters": [],
                "responses": {
                    "default": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "oauth/v1/forgot-pw-reset": {
            "post": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "example": "any"
                                    },
                                    "token": {
                                        "example": "any"
                                    },
                                    "tempPassword": {
                                        "example": "any"
                                    },
                                    "newPassword": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}